refactor: package 구조 수정 - database 클래스가 구현에 맞는 곳에 위치한다. (자바 인메모리 = java)
