+++ src/main/java/codesquad/application/handler/LoginHandler.java
@@ -2,7 +2,7 @@ package codesquad.application.handler;
 
 import codesquad.application.dao.UserDao;
 import codesquad.application.domain.User;
-import codesquad.database.UserDatabase;
+import codesquad.database.java.UserDatabase;
 import codesquad.webserver.annotation.RequestMapping;
 import codesquad.webserver.authentication.AuthenticationHolder;
 import codesquad.webserver.annotation.Controller;
