+++ src/main/java/codesquad/database/java/UserDatabase.java
@@ -1,4 +1,4 @@
-package codesquad.database;
+package codesquad.database.java;
 
 import codesquad.application.dao.UserDao;
 import codesquad.application.domain.User;
