+++ src/test/java/codesquad/application/HtmlPageHandlerTest.java
@@ -3,7 +3,7 @@ package codesquad.application;
 import codesquad.application.dao.UserDao;
 import codesquad.application.handler.HtmlPageHandler;
 import codesquad.application.domain.User;
-import codesquad.database.UserDatabase;
+import codesquad.database.java.UserDatabase;
 import codesquad.webserver.authentication.AuthenticationHolder;
 import codesquad.webserver.http.HttpRequest;
 import codesquad.webserver.http.HttpResponse;
