+++ src/test/java/codesquad/application/LoginHandlerTest.java
@@ -6,8 +6,8 @@ import codesquad.application.dao.SessionDao;
 import codesquad.application.dao.UserDao;
 import codesquad.application.handler.LoginHandler;
 import codesquad.application.domain.User;
-import codesquad.database.SessionDatabase;
-import codesquad.database.UserDatabase;
+import codesquad.database.java.SessionDatabase;
+import codesquad.database.java.UserDatabase;
 import codesquad.webserver.authentication.AuthenticationHolder;
 import codesquad.webserver.http.HttpRequest;
 import codesquad.webserver.http.HttpResponse;
